+++ pedl_train.py
@@ -4,7 +4,7 @@ based on PyTorch to PEDL client.
 """
 import argparse
 import pathlib
-import warnings
+
 
 from typing import Any, Dict, Sequence, Tuple, Union
 
@@ -14,8 +14,9 @@ from torch import nn
 from utils.datasets import LoadImagesAndLabels
 from models import Darknet, attempt_download
 from utils.utils import labels_to_class_weights, compute_loss
+from utils.warmuplr import WarmupMultiStepLR
 import torch.optim as optim
-from torch.optim.lr_scheduler import _LRScheduler
+
 
 import pedl
 from pedl.frameworks.pytorch import PyTorchTrial, LRScheduler
@@ -23,15 +24,8 @@ from pedl.frameworks.pytorch.data import DataLoader
 
 import test
 
-from collections import Counter
 
 TorchData = Union[Dict[str, torch.Tensor], Sequence[torch.Tensor], torch.Tensor]
-
-
-def download_data() -> str:
-    # For a start we hard_code the dataset from bind_mounts
-    # Currently, PEDL Client does not support a download_data_fn
-    return "data"
 
 
 class LazyModule(object):
@@ -96,95 +90,6 @@ def make_data_loaders(
     )
 
 
-class WarmupMultiStepLR(_LRScheduler):
-    """Decays the learning rate of each parameter group by gamma once the
-    number of epoch reaches one of the milestones. Notice that such decay can
-    happen simultaneously with other changes to the learning rate from outside
-    this scheduler. When last_epoch=-1, sets initial lr as lr.
-
-    Warmup part, sets the i-th component of lr to an early value for the first few
-    epochs.
-
-    Args:
-        optimizer (Optimizer): Wrapped optimizer.
-        milestones (list): List of epoch indices. Must be increasing.
-        gamma (float): Multiplicative factor of learning rate decay.
-            Default: 0.1.
-        last_epoch (int): The index of last epoch. Default: -1.
-        threshold (int): The number of epochs to use the lower value
-            Default: 3
-        early_value (float): The value the learning rate should be set to for early epochs
-            Default: 0.1
-        index (int): The index of the learning rate array that should be changed. Changes all
-                    values if index is set to -1.
-            Default: -1
-
-    Example:
-        >>> # Assuming optimizer uses lr = 0.05 for all groups
-        >>> # lr = 0.1      if epoch < 3
-        >>> # lr = 0.05     if 3<= epoch < 30
-        >>> # lr = 0.005    if 30 <= epoch < 80
-        >>> # lr = 0.0005   if epoch >= 80
-        >>> scheduler = WarmupMultiStepLR(optimizer, milestones=[30,80], gamma=0.1)
-        >>> for epoch in range(100):
-        >>>     train(...)
-        >>>     validate(...)
-        >>>     scheduler.step()
-    """
-
-    def __init__(
-        self,
-        optimizer,
-        milestones,
-        gamma=0.1,
-        last_epoch=-1,
-        threshold=3,
-        early_value=0.1,
-        index=-1,
-    ):
-        self.milestones = Counter(milestones)
-        self.gamma = gamma
-        self.threshold = threshold
-        self.early_value = early_value
-        self.index = index
-        self.original_value = 0.0
-        self.memory_set = False
-        super(WarmupMultiStepLR, self).__init__(optimizer, last_epoch)
-
-    def get_lr(self):
-        if not self._get_lr_called_within_step:
-            warnings.warn(
-                "To get the last learning rate computed by the scheduler, "
-                "please use `get_last_lr()`.",
-                DeprecationWarning,
-            )
-        if self.last_epoch < self.threshold:
-            values = [group["lr"] for group in self.optimizer.param_groups]
-            if not self.memory_set:
-                if self.index == -1:
-                    self.original_value = values
-                else:
-                    self.original_value = values[self.index]
-                self.memory_set = True
-            values[self.index] = self.early_value
-            return values
-
-        if self.last_epoch == self.threshold:
-            values = [group["lr"] for group in self.optimizer.param_groups]
-            if self.index == -1:
-                values = self.original_value
-            else:
-                values[self.index] = self.original_value
-            return values
-
-        if self.last_epoch not in self.milestones:
-            return [group["lr"] for group in self.optimizer.param_groups]
-        return [
-            group["lr"] * self.gamma ** self.milestones[self.last_epoch]
-            for group in self.optimizer.param_groups
-        ]
-
-
 class YOLOv3Trial(PyTorchTrial):
     def build_model(self) -> nn.Module:
 
@@ -192,6 +97,7 @@ class YOLOv3Trial(PyTorchTrial):
         model = Darknet(cfg, arc=opt.arc)  # .to(device)
 
         # Fetch starting weights
+        # TODO Once download_data_fn is implemented this should go into download_data
         attempt_download(weights)
         chkpt = torch.load(weights)
         # load model
@@ -241,7 +147,7 @@ class YOLOv3Trial(PyTorchTrial):
             early_value=0.1,
         )
 
-        return LRScheduler(scheduler, step_every_epoch=True,)
+        return LRScheduler(scheduler, step_every_epoch=True)
 
     def optimizer(self, model: nn.Module) -> torch.optim.Optimizer:  # type: ignore
 
@@ -302,17 +208,9 @@ class YOLOv3Trial(PyTorchTrial):
             single_cls=opt.single_cls,
             dataloader=data_loader,
         )
-        keys = [
-            "P",
-            "R",
-            "mAP_at_0.5",
-            "F1",
-            "Some_number_1",
-            "Some_number_2",
-            "Some_number_3",
-        ]
+        keys = ["P", "R", "mAP_at_0.5", "F1"]
 
-        return dict(zip(keys[:4], results))
+        return dict(zip(keys, results))
 
 
 if __name__ == "__main__":
@@ -401,7 +299,7 @@ if __name__ == "__main__":
         "searcher": {
             "name": "single",
             "metric": "mAP_at_0.5",
-            "max_steps": int(epochs * images * batch_size * accumulate / 100),
+            "max_steps": int(epochs * images / batch_size * accumulate / 100),
             "smaller_is_better": False,
         },
         "max_restarts": 0,
@@ -410,8 +308,13 @@ if __name__ == "__main__":
         ],
         "optimizations": {"aggregation_frequency": accumulate},
         "min_validation_period": int(
-            images * batch_size * accumulate / 100 / 4
+            images / batch_size * accumulate / 100 / 4
         ),  # validate after each 1/4 epoch
+        "resources": {
+            "slots_per_trial": 1,
+            # "distributed": False,
+            # "optimized_parallel": False,
+        },
     }
 
     exp = pedl.Experiment(
